add a rack middleware for request stats-ing
still working out how exactly to deal with namespaces. the imagined use
case is to aggregate metrics on a per-host and per-service basis, so
that if host1 is better/worse that others (code update, config change,
hw issue, etc.) you can see it individually, and also track the service
as a whole (which you obviously want). A better API to support this is
desirable.